Clean-up code and create file
