+++ libs/shared/devops-copilot/feature/src/lib/hooks/use-qovery-context/use-qovery-context.ts
@@ -1,0 +1,60 @@
+import { useMatch, useParams } from 'react-router-dom'
+import { useCluster } from '@qovery/domains/clusters/feature'
+import { useEnvironment } from '@qovery/domains/environments/feature'
+import { useOrganization } from '@qovery/domains/organizations/feature'
+import { useProject } from '@qovery/domains/projects/feature'
+import { useService } from '@qovery/domains/services/feature'
+import { DEPLOYMENT_LOGS_VERSION_URL, ENVIRONMENT_LOGS_URL, SERVICE_LOGS_URL } from '@qovery/shared/routes'
+
+export function useQoveryContext() {
+  const {
+    organizationId = '',
+    clusterId = '',
+    projectId = '',
+    environmentId,
+    databaseId,
+    applicationId,
+    serviceId,
+  } = useParams()
+
+  const matchServiceLogs = useMatch({ path: ENVIRONMENT_LOGS_URL() + SERVICE_LOGS_URL(), end: false })?.params
+  const matchDeploymentLogs = useMatch({
+    path: ENVIRONMENT_LOGS_URL() + DEPLOYMENT_LOGS_VERSION_URL(),
+    end: false,
+  })?.params
+
+  const _serviceId =
+    applicationId || serviceId || databaseId || matchServiceLogs?.['serviceId'] || matchDeploymentLogs?.['serviceId']
+
+  const { data: organization } = useOrganization({ organizationId })
+  const { data: project } = useProject({ organizationId, projectId })
+  const { data: environment } = useEnvironment({ environmentId })
+  const { data: cluster } = useCluster({ organizationId, clusterId: clusterId ?? environment?.cluster_id })
+  const { data: service } = useService({ environmentId, serviceId: _serviceId })
+
+  const entityMap = [
+    [service, 'service'],
+    [environment, 'environment'],
+    [project, 'project'],
+    [cluster, 'cluster'],
+    [organization, 'organization'],
+  ]
+
+  const current = entityMap.find(([entity]) => entity !== undefined)
+
+  return {
+    context: {
+      organization,
+      cluster,
+      project,
+      environment,
+      service,
+      deployment: service && {
+        execution_id: matchDeploymentLogs?.['versionId'],
+      },
+    },
+    current: current && typeof current[0] === 'object' ? { ...current[0], type: current[1] } : undefined,
+  }
+}
+
+export default useQoveryContext
